feat: derive Copy and Clone for Qualifier (#69)
I ran into a problem, where I needed to copy the `qual` field from an
ACLEntry instance (which is borrowed), but compiler didn't allow me to
do this, because the Qualifier enum is not copyable, and in my case move
isn't possible and desirable. Deriving the Copy and Clone traits fixes
this.